add SweetAlert

+++ src/app/task-detail/task-detail.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { Task } from 'src/model/task';
 import { ApiService } from 'src/service/api.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-task-detail',
@@ -19,14 +20,27 @@ export class TaskDetailComponent implements OnInit {
 
   getTask(id) {
     this.api.getTask(id)
-      .subscribe(sata => {
+      .subscribe(data => {
+        this.task = data;
+        console.log(this.task);
         this.isLoadingresults = false;
       });
   }
 
   deleteTask(id) {
-    this.isLoadingresults = true;
-    this.api.deleteTask(id)
+
+    Swal.fire({
+      title: 'Está certo disso?',
+      text: 'você não poderá reverter essa operação',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Sim, vala!'
+    }).then((result) => {
+      if (result.value) {
+        this.api.deleteTask(id)
       .subscribe(res => {
         this.isLoadingresults = false;
         this.router.navigate(['/task']);
@@ -34,6 +48,14 @@ export class TaskDetailComponent implements OnInit {
         console.log(err);
         this.isLoadingresults = false;
       });
+        Swal.fire(
+          'Vala!',
+          'Sua tarefa foi apagada!',
+          'success'
+        );
+      }
+    });
+
   }
 
 }
